requirements.txt
mysql.connector
pandas
pygame